TaskList now takes properties
Pain.
In more detail, you can now pass an array of data for tasks into TaskList, and TaskList will pass each spot in the array to Task, and then you'll get a bunch of tasks back in your page. There was an annoying bug with tasks though, where you have to type tasks.tasks.(whatever) because for some reason the code passes tasks, which contains the object tasks, which contains the array. I hate JavaScript.