Description / Date fix
Rebuilt structure so that task.description is never saved as an Array, rather, it is now split into an array only for display.

Also got Date to work right - as far as I can tell. Display now matches Date Picker format, and now auto-populates the date when editing and saves correctly! Yay!