Cosmetic/UI Upgrade.
I turned all of the tasks into accordions that show the description when they drop down. I gave every task its own edit and delete button. These buttons are modified versions of the Task edit and delete button. I made it so the edit button pops up a modal that you use to edit the task. I also changed the delete button to says finished because it makes since for it to be removed when finished. I turned the login system into a modal that pops up whenever you first load the page. I just took all of the code for the login system and put it in a modal. I got rid of the TaskForm component and replaced it with NewTaskButton that stays at the top right of the screen as you scroll to make it easier to make new tasks. This opens a modal that allows you to create a new task. I took the color scheme and button designs that Oli made and put them on the buttons and modals. I think that was everything.

P.S.
The application now uses react-bootstrap so you will now also need to npm install react-bootstrap inorder to run the application.